removed dead backwards compatible code (for API level < 10) changed hard coded integer references to API levels to use platform provided constants.
